Rename index.ts to show-hide.ts
